14) Smart Health Assistant/requirements.txt
# requirements.txt
pyautogen==0.2.28
streamlit==1.36.0
google-generativeai==0.5.4
streamlit-extras